+++ internal/app/wwctl/overlay/create/root.go
@@ -2,6 +2,7 @@ package create
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/warewulf/warewulf/internal/app/wwctl/completions"
 )
 
 var (
@@ -13,6 +14,7 @@ var (
 		Long:                  "This command creates a new empty overlay with the given OVERLAY_NAME.",
 		RunE:                  CobraRunE,
 		Args:                  cobra.ExactArgs(1),
+		ValidArgsFunction:     completions.None,
 	}
 )